+++ Gemfile
@@ -2,5 +2,6 @@ gem 'fiscali'
 
 group :test do
   gem 'rails-controller-testing'
-  gem 'byebug'
+  #gem 'byebug'
+  gem 'byebug', '~>11.1.3'
 end
